requirements.txt
Django==1.10.3
Markdown==2.6.7
PyJWT==0.4.1
altgraph==0.10.2
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
bdist-mpkg==0.5.0
bonjour-py==0.3
boto==2.42.0
braintree==3.10.0
certifi==2015.11.20.1
dj-database-url==0.4.1
dj-static==0.0.6
django-appconf==0.6
django-compressor==1.4
django-crispy-forms==1.4.0
django-filter==0.15.3
django-storages==1.5.1
djangorestframework==3.5.2
drf-nested-routers==0.9.0
googlemaps==2.4.4
gunicorn==19.6.0
httplib2==0.9.2
macholib==1.5.1
matplotlib==1.3.1
modulegraph==0.10.4
nose==1.3.7
numpy==1.8.0rc1
oauth2==1.9.0.post1
oauthlib==0.7.2
pandas==0.17.1
pbr==1.8.1
py2app==0.7.3
pyOpenSSL==0.13.1
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyparsing==2.0.1
python-dateutil==1.5
python-openid==2.2.5
python-social-auth==0.2.1
pytz==2013.7
requests==2.10.0
requests-oauthlib==0.4.2
scikit-learn==0.17
scipy==0.13.0b1
singledispatch==3.4.0.3
six==1.4.1
skyscanner==1.1.4
static3==0.5.1
stevedore==1.12.0
tornado==4.3
vboxapi==1.0
virtualenv==14.0.6
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
wsgiref==0.1.2
xattr==0.6.4
yelp==1.0.2
zope.interface==4.1.1
